Changed doubles to ints, reformatted some --Braden

+++ src/main/java/frc/robot/subsystems/Node.java
@@ -21,8 +21,8 @@ public class Node {
 
   int x;
   int y;
-  double hValue;
+  int hValue;
   int gValue;
-  double fValue;
+  int fValue;
   Node parent;
 }
